Address CodeRabbit review: locale-safe printf; drop redundant int()
- Major (statusline-snippet.sh): bash printf %.0f is locale-aware, so on a
  comma-decimal locale (de_DE, fr_FR) it errors on the JSON float "11.5"
  and yields empty output plus stderr spam every render. Force LC_NUMERIC=C
  for that printf so the dot decimal always parses.
- Trivial (usage.py): round() already returns int in Python 3; drop the
  redundant int() cast in _bar.

+++ usage.py
@@ -538,7 +538,7 @@ def _bar(pct, cells=10):
         p = max(0.0, min(100.0, float(pct)))
     except (TypeError, ValueError):
         p = 0.0
-    filled = max(0, min(cells, int(round(p / 100 * cells))))
+    filled = max(0, min(cells, round(p / 100 * cells)))
     return "[" + "█" * filled + "░" * (cells - filled) + "]"
